refactor(seq): move get_common_length from commands/ancestral/fitch to seq/alignment
- Alignment validation utility, not an ancestral reconstruction algorithm
- Used by 8 production files across 4 domains (ancestral, optimize, timetree, representation)
- All import paths updated to crate::seq::alignment::get_common_length
