+++ packages/treetime/src/commands/ancestral/__tests__/test_marginal_analytical/test_marginal_analytical_support.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 pub mod tests {
   use crate::alphabet::alphabet::{Alphabet, AlphabetName};
-  use crate::commands::ancestral::fitch::get_common_length;
+  use crate::seq::alignment::get_common_length;
   use crate::commands::ancestral::marginal::initialize_marginal;
   use crate::gtr::gtr::GTR;
   use crate::representation::partition::marginal_dense::PartitionMarginalDense;
