+++ packages/treetime/src/commands/optimize/__tests__/test_initial_guess_gtr_messages.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests {
   use crate::alphabet::alphabet::Alphabet;
-  use crate::commands::ancestral::fitch::get_common_length;
+  use crate::seq::alignment::get_common_length;
   use crate::commands::ancestral::marginal::{initialize_marginal, update_marginal};
   use crate::commands::optimize::optimize_unified::initial_guess_mixed;
   use crate::gtr::get_gtr::{F81Params, JC69Params, f81, jc69};
