+++ packages/treetime/src/representation/partition/fitch_config.rs
@@ -1,5 +1,5 @@
 use crate::alphabet::alphabet::Alphabet;
-use crate::commands::ancestral::fitch::get_common_length;
+use crate::seq::alignment::get_common_length;
 use eyre::Report;
 use treetime_io::fasta::FastaRecord;
