+++ packages/treetime/src/test_utils.rs
@@ -1,5 +1,5 @@
 use crate::alphabet::alphabet::{Alphabet, AlphabetName};
-use crate::commands::ancestral::fitch::get_common_length;
+use crate::seq::alignment::get_common_length;
 use crate::commands::ancestral::marginal::{initialize_marginal, update_marginal};
 use crate::gtr::gtr::GTR;
 use crate::representation::partition::fitch::PartitionFitch;
